Adjust verification tests for header-only reporting
