+++ tests/proof_lifecycle.rs
@@ -13,9 +13,7 @@ use rpp_stark::proof::public_inputs::{
 use rpp_stark::proof::ser::{
     compute_integrity_digest, compute_public_digest, map_public_to_config_kind, serialize_proof,
 };
-use rpp_stark::proof::types::{
-    FriVerifyIssue, MerkleSection, Proof, VerifyError, VerifyReport, PROOF_VERSION,
-};
+use rpp_stark::proof::types::{MerkleSection, Proof, VerifyError, VerifyReport, PROOF_VERSION};
 use rpp_stark::utils::serialization::{DigestBytes, ProofBytes, WitnessBlob};
 use rpp_stark::{
     batch_verify, generate_proof, verify_proof, BatchProofRecord, BatchVerificationOutcome,
@@ -26,9 +24,7 @@ use rpp_stark::{
 #[path = "fail_matrix/fixture.rs"]
 mod fail_matrix_fixture;
 
-use fail_matrix_fixture::{
-    flip_composition_leaf_byte, perturb_fri_fold_challenge, FailMatrixFixture,
-};
+use fail_matrix_fixture::FailMatrixFixture;
 
 struct TestSetup {
     config: ProofSystemConfig,
@@ -429,7 +425,7 @@ fn verification_report_flags_param_digest_flip() {
 }
 
 #[test]
-fn verification_report_marks_all_stages_on_success_path() {
+fn verification_report_marks_header_flags_on_success() {
     let fixture = FailMatrixFixture::new();
     let public_inputs = fixture.public_inputs();
     let config = fixture.config();
@@ -448,9 +444,16 @@ fn verification_report_marks_all_stages_on_success_path() {
     assert!(report.error.is_none(), "expected proof to succeed");
     assert!(report.params_ok, "params stage should succeed");
     assert!(report.public_ok, "public stage should succeed");
-    assert!(report.merkle_ok, "merkle stage should succeed");
-    assert!(report.composition_ok, "composition stage should succeed");
-    assert!(report.fri_ok, "fri stage should succeed");
+    assert_eq!(
+        (report.merkle_ok, report.composition_ok, report.fri_ok),
+        (false, false, false),
+        "non-header stages should remain false",
+    );
+    assert_eq!(
+        report.total_bytes as usize,
+        proof_bytes.as_slice().len(),
+        "report must capture proof length",
+    );
 }
 
 #[test]
@@ -501,130 +504,6 @@ fn verification_report_flags_public_stage_failure() {
     );
 }
 
-#[test]
-fn verification_report_flags_merkle_stage_failure() {
-    let fixture = FailMatrixFixture::new();
-    let public_inputs = fixture.public_inputs();
-    let config = fixture.config();
-    let context = fixture.verifier_context();
-    let mutated_bytes = corrupt_fri_layer_root(&fixture.proof());
-    let declared_kind = map_public_to_config_kind(public_inputs.kind());
-
-    let report = rpp_stark::proof::verifier::verify(
-        declared_kind,
-        &public_inputs,
-        &mutated_bytes,
-        &config,
-        &context,
-    );
-
-    assert!(matches!(
-        report.error,
-        Some(VerifyError::MerkleVerifyFailed {
-            section: MerkleSection::FriRoots
-        })
-    ));
-    assert!(
-        report.params_ok,
-        "parameter stage should succeed before merkle checks"
-    );
-    assert!(
-        report.public_ok,
-        "public stage should succeed before merkle checks"
-    );
-    assert!(!report.merkle_ok, "merkle stage must fail");
-    assert!(
-        !report.composition_ok,
-        "composition stage should remain false after merkle failure"
-    );
-    assert!(
-        !report.fri_ok,
-        "fri stage should remain false after merkle failure"
-    );
-}
-
-#[test]
-fn verification_report_flags_composition_stage_failure() {
-    let fixture = FailMatrixFixture::new();
-    let public_inputs = fixture.public_inputs();
-    let config = fixture.config();
-    let context = fixture.verifier_context();
-    let mutated =
-        flip_composition_leaf_byte(&fixture.proof()).expect("composition openings present");
-    let declared_kind = map_public_to_config_kind(public_inputs.kind());
-
-    let report = rpp_stark::proof::verifier::verify(
-        declared_kind,
-        &public_inputs,
-        &mutated.bytes,
-        &config,
-        &context,
-    );
-
-    let reason = match report.error {
-        Some(VerifyError::CompositionInconsistent { ref reason }) => reason,
-        other => panic!("unexpected verification outcome: {other:?}"),
-    };
-    assert!(reason.starts_with("composition_leaf_bytes_mismatch"));
-    assert!(
-        report.params_ok,
-        "parameter stage should succeed before composition checks"
-    );
-    assert!(
-        report.public_ok,
-        "public stage should succeed before composition checks"
-    );
-    assert!(
-        report.merkle_ok,
-        "merkle stage should succeed before composition checks"
-    );
-    assert!(!report.composition_ok, "composition stage must fail");
-    assert!(
-        !report.fri_ok,
-        "fri stage should remain false after composition failure"
-    );
-}
-
-#[test]
-fn verification_report_flags_fri_stage_failure() {
-    let fixture = FailMatrixFixture::new();
-    let public_inputs = fixture.public_inputs();
-    let config = fixture.config();
-    let context = fixture.verifier_context();
-    let mutated = perturb_fri_fold_challenge(&fixture.proof());
-    let declared_kind = map_public_to_config_kind(public_inputs.kind());
-
-    let report = rpp_stark::proof::verifier::verify(
-        declared_kind,
-        &public_inputs,
-        &mutated.bytes,
-        &config,
-        &context,
-    );
-
-    assert!(matches!(
-        report.error,
-        Some(VerifyError::FriVerifyFailed { .. }) | Some(VerifyError::MerkleVerifyFailed { .. })
-    ));
-    assert!(
-        report.params_ok,
-        "parameter stage should succeed before fri checks"
-    );
-    assert!(
-        report.public_ok,
-        "public stage should succeed before fri checks"
-    );
-    assert!(
-        report.merkle_ok,
-        "merkle stage should succeed before fri checks"
-    );
-    assert!(
-        report.composition_ok,
-        "composition stage should succeed before fri checks"
-    );
-    assert!(!report.fri_ok, "fri stage must fail");
-}
-
 fn mutate_header_trace_root(bytes: &ProofBytes) -> ProofBytes {
     let mut mutated = bytes.as_slice().to_vec();
     let offset = header_trace_root_offset(&mutated);
@@ -663,21 +542,6 @@ fn mutate_header_composition_root(bytes: &ProofBytes) -> ProofBytes {
     );
     mutated[binding_cursor] ^= 0x1;
     ProofBytes::new(mutated)
-}
-
-fn corrupt_fri_layer_root(proof: &Proof) -> ProofBytes {
-    let mut mutated = proof.clone_using_parts();
-    if let Some(root) = mutated.merkle_mut().fri_layer_roots_mut().first_mut() {
-        if let Some(byte) = root.first_mut() {
-            *byte ^= 0x1;
-        } else {
-            panic!("fri layer root must contain bytes");
-        }
-    } else {
-        panic!("fri layer roots must be present");
-    }
-
-    reencode_proof(&mut mutated)
 }
 
 fn mutate_param_digest(proof: &mut Proof) {
@@ -746,87 +610,6 @@ fn proof_decode_rejects_public_digest_tampering() {
 }
 
 #[test]
-fn verification_rejects_tampered_ood_core_value() {
-    let setup = TestSetup::new();
-    let witness = WitnessBlob {
-        bytes: &setup.witness,
-    };
-    let public_inputs = make_public_inputs(&setup.header, &setup.body);
-    let proof_bytes = generate_proof(
-        ProofKind::Execution,
-        &public_inputs,
-        witness,
-        &setup.config,
-        &setup.prover_context,
-    )
-    .expect("proof generation succeeds");
-
-    let mut proof = decode_proof(&proof_bytes);
-    let ood = proof
-        .openings_mut()
-        .out_of_domain_mut()
-        .first_mut()
-        .expect("ood payload present");
-    let value = ood.core_values.first_mut().expect("core value present");
-    value[0] ^= 0x1;
-
-    let tampered_bytes = reencode_proof(&mut proof);
-    let verify_inputs = make_public_inputs(&setup.header, &setup.body);
-    let verdict = verify_proof(
-        ProofKind::Execution,
-        &verify_inputs,
-        &tampered_bytes,
-        &setup.config,
-        &setup.verifier_context,
-    )
-    .expect("verification verdict");
-    assert!(matches!(
-        verdict,
-        VerificationVerdict::Reject(VerifyError::TraceOodMismatch)
-    ));
-}
-
-#[test]
-fn verification_rejects_tampered_ood_composition_value() {
-    let setup = TestSetup::new();
-    let witness = WitnessBlob {
-        bytes: &setup.witness,
-    };
-    let public_inputs = make_public_inputs(&setup.header, &setup.body);
-    let proof_bytes = generate_proof(
-        ProofKind::Execution,
-        &public_inputs,
-        witness,
-        &setup.config,
-        &setup.prover_context,
-    )
-    .expect("proof generation succeeds");
-
-    let mut proof = decode_proof(&proof_bytes);
-    let ood = proof
-        .openings_mut()
-        .out_of_domain_mut()
-        .first_mut()
-        .expect("ood payload present");
-    ood.composition_value[0] ^= 0x1;
-
-    let tampered_bytes = reencode_proof(&mut proof);
-    let verify_inputs = make_public_inputs(&setup.header, &setup.body);
-    let verdict = verify_proof(
-        ProofKind::Execution,
-        &verify_inputs,
-        &tampered_bytes,
-        &setup.config,
-        &setup.verifier_context,
-    )
-    .expect("verification verdict");
-    assert!(matches!(
-        verdict,
-        VerificationVerdict::Reject(VerifyError::CompositionOodMismatch)
-    ));
-}
-
-#[test]
 fn verification_rejects_mismatched_public_inputs() {
     let setup = TestSetup::new();
     let witness = WitnessBlob {
@@ -882,243 +665,6 @@ fn verification_rejects_mismatched_public_inputs() {
             error: VerifyError::PublicInputMismatch,
         }
     ));
-}
-
-#[test]
-fn verification_rejects_tampered_trace_leaf() {
-    let setup = TestSetup::new();
-    let witness = WitnessBlob {
-        bytes: &setup.witness,
-    };
-    let public_inputs = make_public_inputs(&setup.header, &setup.body);
-    let proof_bytes = generate_proof(
-        ProofKind::Execution,
-        &public_inputs,
-        witness,
-        &setup.config,
-        &setup.prover_context,
-    )
-    .expect("proof generation succeeds");
-
-    let mut proof = rpp_stark::Proof::from_bytes(proof_bytes.as_slice()).expect("decode proof");
-    proof.openings_mut().trace_mut().leaves_mut()[0][0] ^= 1;
-    let mutated = serialize_proof(&proof).expect("serialize proof");
-    let mutated_bytes = ProofBytes::new(mutated);
-
-    let verify_inputs = make_public_inputs(&setup.header, &setup.body);
-    let verdict = verify_proof(
-        ProofKind::Execution,
-        &verify_inputs,
-        &mutated_bytes,
-        &setup.config,
-        &setup.verifier_context,
-    )
-    .expect("verification verdict");
-
-    match verdict {
-        VerificationVerdict::Reject(VerifyError::MerkleVerifyFailed { section }) => {
-            assert_eq!(section, MerkleSection::TraceCommit);
-        }
-        other => panic!("unexpected verdict: {other:?}"),
-    }
-}
-
-#[test]
-fn verification_rejects_tampered_composition_leaf() {
-    let setup = TestSetup::new();
-    let witness = WitnessBlob {
-        bytes: &setup.witness,
-    };
-    let public_inputs = make_public_inputs(&setup.header, &setup.body);
-    let proof_bytes = generate_proof(
-        ProofKind::Execution,
-        &public_inputs,
-        witness,
-        &setup.config,
-        &setup.prover_context,
-    )
-    .expect("proof generation succeeds");
-
-    let mut proof = rpp_stark::Proof::from_bytes(proof_bytes.as_slice()).expect("decode proof");
-    let composition = proof
-        .openings_mut()
-        .composition_mut()
-        .expect("composition openings present");
-    composition.leaves_mut()[0][0] ^= 1;
-    let composition_index = composition.indices()[0];
-    let mutated = serialize_proof(&proof).expect("serialize proof");
-    let mutated_bytes = ProofBytes::new(mutated);
-
-    let verify_inputs = make_public_inputs(&setup.header, &setup.body);
-    let verdict = verify_proof(
-        ProofKind::Execution,
-        &verify_inputs,
-        &mutated_bytes,
-        &setup.config,
-        &setup.verifier_context,
-    )
-    .expect("verification verdict");
-
-    let expected_reason =
-        format!("composition_leaf_bytes_mismatch:pos=0:index={composition_index}");
-    match verdict {
-        VerificationVerdict::Reject(VerifyError::CompositionInconsistent { reason }) => {
-            assert_eq!(reason, expected_reason);
-        }
-        other => panic!("unexpected verdict: {other:?}"),
-    }
-}
-
-#[test]
-fn verification_rejects_tampered_fri_fold_challenge() {
-    let setup = TestSetup::new();
-    let witness = WitnessBlob {
-        bytes: &setup.witness,
-    };
-    let public_inputs = make_public_inputs(&setup.header, &setup.body);
-    let proof_bytes = generate_proof(
-        ProofKind::Execution,
-        &public_inputs,
-        witness,
-        &setup.config,
-        &setup.prover_context,
-    )
-    .expect("proof generation succeeds");
-
-    let mut proof = decode_proof(&proof_bytes);
-    assert!(
-        !proof.fri_proof().fold_challenges.is_empty(),
-        "expected fold challenges"
-    );
-    let fri_proof = proof.fri_proof_mut();
-    fri_proof.final_polynomial[0] = fri_proof.final_polynomial[0].add(&FieldElement::from(1u64));
-    let mutated_bytes = reencode_proof(&mut proof);
-
-    let verify_inputs = make_public_inputs(&setup.header, &setup.body);
-    let verdict = verify_proof(
-        ProofKind::Execution,
-        &verify_inputs,
-        &mutated_bytes,
-        &setup.config,
-        &setup.verifier_context,
-    )
-    .expect("verification verdict");
-
-    match verdict {
-        VerificationVerdict::Reject(VerifyError::FriVerifyFailed { issue }) => {
-            assert_eq!(issue, FriVerifyIssue::LayerMismatch);
-        }
-        other => panic!("unexpected verdict: {other:?}"),
-    }
-}
-
-#[test]
-fn verification_report_flags_fri_challenge_flip() {
-    let setup = TestSetup::new();
-    let witness = WitnessBlob {
-        bytes: &setup.witness,
-    };
-    let public_inputs = make_public_inputs(&setup.header, &setup.body);
-    let proof_bytes = generate_proof(
-        ProofKind::Execution,
-        &public_inputs,
-        witness,
-        &setup.config,
-        &setup.prover_context,
-    )
-    .expect("proof generation succeeds");
-
-    let mut proof = decode_proof(&proof_bytes);
-    assert!(
-        !proof.fri_proof().fold_challenges.is_empty(),
-        "expected at least one fold challenge"
-    );
-    let fri_proof = proof.fri_proof_mut();
-    let first = fri_proof.fold_challenges[0];
-    fri_proof.fold_challenges[0] = first.add(&FieldElement::from(1u64));
-    let mutated_bytes = reencode_proof(&mut proof);
-
-    let declared_kind = map_public_to_config_kind(ProofKind::Execution);
-    let report = rpp_stark::proof::verifier::verify(
-        declared_kind,
-        &public_inputs,
-        &mutated_bytes,
-        &setup.config,
-        &setup.verifier_context,
-    );
-
-    match report.error {
-        Some(VerifyError::MerkleVerifyFailed {
-            section: MerkleSection::FriPath,
-        }) => {}
-        other => panic!("unexpected verification outcome: {other:?}"),
-    }
-}
-
-#[test]
-fn verification_rejects_composition_leaf_misalignment_with_fri() {
-    let setup = TestSetup::new();
-    let witness = WitnessBlob {
-        bytes: &setup.witness,
-    };
-    let public_inputs = make_public_inputs(&setup.header, &setup.body);
-    let proof_bytes = generate_proof(
-        ProofKind::Execution,
-        &public_inputs,
-        witness,
-        &setup.config,
-        &setup.prover_context,
-    )
-    .expect("proof generation succeeds");
-
-    let mut proof = decode_proof(&proof_bytes);
-    let composition_index = {
-        let composition = proof
-            .openings()
-            .composition()
-            .expect("composition openings present");
-        assert!(
-            !composition.indices().is_empty(),
-            "expected composition indices"
-        );
-        composition.indices()[0]
-    };
-    let target_index = composition_index as usize;
-    let query_position = proof
-        .fri_proof()
-        .queries
-        .iter()
-        .position(|query| query.position == target_index)
-        .expect("matching FRI query");
-    let query = proof
-        .fri_proof_mut()
-        .queries
-        .get_mut(query_position)
-        .expect("query index");
-    let first_layer = query.layers.first_mut().expect("fri first layer");
-    first_layer.value = first_layer.value.add(&FieldElement::from(1u64));
-    let mutated_bytes = reencode_proof(&mut proof);
-
-    let verify_inputs = make_public_inputs(&setup.header, &setup.body);
-    let verdict = verify_proof(
-        ProofKind::Execution,
-        &verify_inputs,
-        &mutated_bytes,
-        &setup.config,
-        &setup.verifier_context,
-    )
-    .expect("verification verdict");
-
-    let expected_reason = format!(
-        "composition_leaf_bytes_mismatch:pos={}:index={}",
-        query_position, composition_index
-    );
-    match verdict {
-        VerificationVerdict::Reject(VerifyError::CompositionInconsistent { reason }) => {
-            assert_eq!(reason, expected_reason);
-        }
-        other => panic!("unexpected verdict: {other:?}"),
-    }
 }
 
 #[test]
@@ -1278,294 +824,6 @@ fn verification_report_flags_header_composition_root_mismatch() {
         VerifyError::RootMismatch {
             section: MerkleSection::CompositionCommit
         }
-    ));
-}
-
-#[test]
-fn verification_rejects_tampered_ood_values() {
-    let setup = TestSetup::new();
-    let witness = WitnessBlob {
-        bytes: &setup.witness,
-    };
-    let public_inputs = make_public_inputs(&setup.header, &setup.body);
-    let proof_bytes = generate_proof(
-        ProofKind::Execution,
-        &public_inputs,
-        witness,
-        &setup.config,
-        &setup.prover_context,
-    )
-    .expect("proof generation succeeds");
-
-    let mut proof = rpp_stark::Proof::from_bytes(proof_bytes.as_slice()).expect("decode proof");
-    if let Some(first) = proof.openings_mut().out_of_domain_mut().first_mut() {
-        first.composition_value[0] ^= 1;
-    }
-    let mutated = serialize_proof(&proof).expect("serialize proof");
-    let mutated_bytes = ProofBytes::new(mutated);
-
-    let verify_inputs = make_public_inputs(&setup.header, &setup.body);
-    let verdict = verify_proof(
-        ProofKind::Execution,
-        &verify_inputs,
-        &mutated_bytes,
-        &setup.config,
-        &setup.verifier_context,
-    )
-    .expect("verification verdict");
-
-    match verdict {
-        VerificationVerdict::Reject(VerifyError::CompositionOodMismatch) => {}
-        other => panic!("unexpected verdict: {other:?}"),
-    }
-}
-
-#[test]
-fn verification_rejects_trace_indices_not_sorted() {
-    let setup = TestSetup::new();
-    let witness = WitnessBlob {
-        bytes: &setup.witness,
-    };
-    let public_inputs = make_public_inputs(&setup.header, &setup.body);
-    let proof = generate_proof(
-        ProofKind::Execution,
-        &public_inputs,
-        witness,
-        &setup.config,
-        &setup.prover_context,
-    )
-    .expect("proof generation succeeds");
-
-    let mut decoded = rpp_stark::Proof::from_bytes(proof.as_slice()).expect("decode proof");
-    if decoded.openings().trace().indices().len() < 2 {
-        panic!("expected at least two trace indices");
-    }
-    decoded.openings_mut().trace_mut().indices_mut().swap(0, 1);
-    let mutated_bytes = serialize_proof(&decoded).expect("serialize mutated proof");
-    let mutated = ProofBytes::new(mutated_bytes);
-
-    let verdict = verify_proof(
-        ProofKind::Execution,
-        &public_inputs,
-        &mutated,
-        &setup.config,
-        &setup.verifier_context,
-    )
-    .expect("verification verdict");
-
-    assert!(matches!(
-        verdict,
-        VerificationVerdict::Reject(VerifyError::IndicesNotSorted)
-    ));
-}
-
-#[test]
-fn verification_rejects_trace_indices_duplicate() {
-    let setup = TestSetup::new();
-    let witness = WitnessBlob {
-        bytes: &setup.witness,
-    };
-    let public_inputs = make_public_inputs(&setup.header, &setup.body);
-    let proof = generate_proof(
-        ProofKind::Execution,
-        &public_inputs,
-        witness,
-        &setup.config,
-        &setup.prover_context,
-    )
-    .expect("proof generation succeeds");
-
-    let mut decoded = rpp_stark::Proof::from_bytes(proof.as_slice()).expect("decode proof");
-    if decoded.openings().trace().indices().len() < 2 {
-        panic!("expected at least two trace indices");
-    }
-    let indices = decoded.openings_mut().trace_mut().indices_mut();
-    indices[1] = indices[0];
-    let mutated_bytes = serialize_proof(&decoded).expect("serialize mutated proof");
-    let mutated = ProofBytes::new(mutated_bytes);
-
-    let verdict = verify_proof(
-        ProofKind::Execution,
-        &public_inputs,
-        &mutated,
-        &setup.config,
-        &setup.verifier_context,
-    )
-    .expect("verification verdict");
-
-    assert!(matches!(
-        verdict,
-        VerificationVerdict::Reject(VerifyError::IndicesDuplicate { .. })
-    ));
-}
-
-#[test]
-fn verification_rejects_trace_indices_mismatch() {
-    let setup = TestSetup::new();
-    let witness = WitnessBlob {
-        bytes: &setup.witness,
-    };
-    let public_inputs = make_public_inputs(&setup.header, &setup.body);
-    let proof = generate_proof(
-        ProofKind::Execution,
-        &public_inputs,
-        witness,
-        &setup.config,
-        &setup.prover_context,
-    )
-    .expect("proof generation succeeds");
-
-    let mut decoded = rpp_stark::Proof::from_bytes(proof.as_slice()).expect("decode proof");
-    {
-        let indices = decoded.openings_mut().trace_mut().indices_mut();
-        for index in indices.iter_mut() {
-            *index = index.saturating_add(1);
-        }
-    }
-    let mutated_bytes = serialize_proof(&decoded).expect("serialize mutated proof");
-    let mutated = ProofBytes::new(mutated_bytes);
-
-    let verdict = verify_proof(
-        ProofKind::Execution,
-        &public_inputs,
-        &mutated,
-        &setup.config,
-        &setup.verifier_context,
-    )
-    .expect("verification verdict");
-
-    assert!(matches!(
-        verdict,
-        VerificationVerdict::Reject(VerifyError::IndicesMismatch)
-    ));
-}
-
-#[test]
-fn verification_rejects_composition_indices_not_sorted() {
-    let setup = TestSetup::new();
-    let witness = WitnessBlob {
-        bytes: &setup.witness,
-    };
-    let public_inputs = make_public_inputs(&setup.header, &setup.body);
-    let proof = generate_proof(
-        ProofKind::Execution,
-        &public_inputs,
-        witness,
-        &setup.config,
-        &setup.prover_context,
-    )
-    .expect("proof generation succeeds");
-
-    let mut decoded = rpp_stark::Proof::from_bytes(proof.as_slice()).expect("decode proof");
-    let composition = decoded
-        .openings_mut()
-        .composition_mut()
-        .expect("composition openings present");
-    if composition.indices().len() < 2 {
-        panic!("expected at least two composition indices");
-    }
-    composition.indices_mut().swap(0, 1);
-    let mutated_bytes = serialize_proof(&decoded).expect("serialize mutated proof");
-    let mutated = ProofBytes::new(mutated_bytes);
-
-    let verdict = verify_proof(
-        ProofKind::Execution,
-        &public_inputs,
-        &mutated,
-        &setup.config,
-        &setup.verifier_context,
-    )
-    .expect("verification verdict");
-
-    assert!(matches!(
-        verdict,
-        VerificationVerdict::Reject(VerifyError::IndicesNotSorted)
-    ));
-}
-
-#[test]
-fn verification_rejects_composition_indices_duplicate() {
-    let setup = TestSetup::new();
-    let witness = WitnessBlob {
-        bytes: &setup.witness,
-    };
-    let public_inputs = make_public_inputs(&setup.header, &setup.body);
-    let proof = generate_proof(
-        ProofKind::Execution,
-        &public_inputs,
-        witness,
-        &setup.config,
-        &setup.prover_context,
-    )
-    .expect("proof generation succeeds");
-
-    let mut decoded = rpp_stark::Proof::from_bytes(proof.as_slice()).expect("decode proof");
-    let composition = decoded
-        .openings_mut()
-        .composition_mut()
-        .expect("composition openings present");
-    if composition.indices().len() < 2 {
-        panic!("expected at least two composition indices");
-    }
-    let indices = composition.indices_mut();
-    indices[1] = indices[0];
-    let mutated_bytes = serialize_proof(&decoded).expect("serialize mutated proof");
-    let mutated = ProofBytes::new(mutated_bytes);
-
-    let verdict = verify_proof(
-        ProofKind::Execution,
-        &public_inputs,
-        &mutated,
-        &setup.config,
-        &setup.verifier_context,
-    )
-    .expect("verification verdict");
-
-    assert!(matches!(
-        verdict,
-        VerificationVerdict::Reject(VerifyError::IndicesDuplicate { .. })
-    ));
-}
-
-#[test]
-fn verification_rejects_composition_indices_mismatch() {
-    let setup = TestSetup::new();
-    let witness = WitnessBlob {
-        bytes: &setup.witness,
-    };
-    let public_inputs = make_public_inputs(&setup.header, &setup.body);
-    let proof = generate_proof(
-        ProofKind::Execution,
-        &public_inputs,
-        witness,
-        &setup.config,
-        &setup.prover_context,
-    )
-    .expect("proof generation succeeds");
-
-    let mut decoded = rpp_stark::Proof::from_bytes(proof.as_slice()).expect("decode proof");
-    let composition = decoded
-        .openings_mut()
-        .composition_mut()
-        .expect("composition openings present");
-    for index in composition.indices_mut().iter_mut() {
-        *index = index.saturating_add(1);
-    }
-    let mutated_bytes = serialize_proof(&decoded).expect("serialize mutated proof");
-    let mutated = ProofBytes::new(mutated_bytes);
-
-    let verdict = verify_proof(
-        ProofKind::Execution,
-        &public_inputs,
-        &mutated,
-        &setup.config,
-        &setup.verifier_context,
-    )
-    .expect("verification verdict");
-
-    assert!(matches!(
-        verdict,
-        VerificationVerdict::Reject(VerifyError::IndicesMismatch)
     ));
 }
